+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add function `SKY_coin_BlockHeader_Fee`
 - Add function `SKY_coin_BlockBody_Transactions`
 - Add function `SKY_coin_BlockHeader_UxHash`
 - Add function `SKY_coin_BlockHeader_BkSeq`